HexEdit: Make previous commit compatible with windows platform (poor support for long double :( )
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@5284 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/plugins/contrib/HexEditor/ExpressionPreprocessed.cpp
@@ -122,7 +122,7 @@ namespace Expression
             Value& v = m_CodeArguments[i];
             if ( v.IsFloat() )
             {
-                res += wxString::Format( _T("%d -> Float: %Lf\n"), (int)i, v.GetFloat() );
+                res += wxString::Format( _T("%d -> Float: %f\n"), (int)i, (double)v.GetFloat() );
             }
             else if ( v.IsSignedInt() )
             {
